Fix the git log and git status output processors. Split status field into two.

+++ src/GitHub.Api/Platform/WindowsDiskUsageTask.cs
@@ -3,7 +3,7 @@ using System.Threading;
 
 namespace GitHub.Unity
 {
-    class WindowsDiskUsageTask : ProcessTask<int>
+    class WindowsDiskUsageTask : ProcessTask<long>
     {
         private readonly string arguments;
